requirements.txt
absl-py==1.0.0
aniso8601==9.0.1
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==21.1.0
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
cerebro-dl==1.0.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
diskcache==5.2.1
entrypoints==0.3
filelock==3.4.0
Flask==2.0.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
gast==0.3.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.42.0
h5py==2.10.0
huggingface-hub==0.1.2
hyperopt==0.2.7
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.5.1
ipython==7.29.0
ipython_genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
Keras-Preprocessing==1.1.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.6
numpy==1.18.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
petastorm==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
py4j==0.10.7
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
pyspark==2.4.8
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scipy==1.4.1
Send2Trash==1.8.0
six==1.16.0
SQLAlchemy==1.4.27
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tokenizers==0.10.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.5
typing_extensions==4.0.0
urllib3==1.26.7
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
